requirements.txt
click>=8.1
plotly>=5.18
python-dateutil>=2.8
plotly>=5.18
python-dateutil>=2.8
